requirements.txt
requests
python-dotenv
google-generativeai
PyPDF2
sentence-transformers
torch
